result work

+++ src/main/java/bot/dao/ResultDao.java
@@ -3,7 +3,7 @@ package bot.dao;
 import bot.entity.ResultEntity;
 
 public interface ResultDao {
-    ResultEntity getByKey(Long key);
+    ResultEntity getByKey(Integer key);
     void persist(ResultEntity resultEntity);
     void update(ResultEntity resultEntity);
     void delete(ResultEntity resultEntity);
